+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 Changes:
 
+  * Fixed detection of startDate, containerModel, and indexSequencing for NextSeq 2000 runs
   * Update illumina-interop to 1.2.0
 
 # 1.14.0